requirements.txt
python-dotenv~=1.0.1
sentence_transformers